refactor files

+++ main.py
@@ -1,27 +1,29 @@
 import base64
-import json
 import shutil
 from typing import Optional
-
-from fastapi import FastAPI, File, UploadFile
-from fastapi import Request, Body
-from fastapi.responses import HTMLResponse
-from fastapi.staticfiles import StaticFiles
-from fastapi.templating import Jinja2Templates
+import json
 from pydantic import BaseModel
 from sqlalchemy import and_
 from starlette.middleware.cors import CORSMiddleware
 
-import cart
-import product
+from shop import notes
 import shop
-from cart import cart as cart_db
+from fastapi import Request, Depends, Body
+from fastapi.templating import Jinja2Templates
+from fastapi.staticfiles import StaticFiles
+from fastapi.responses import HTMLResponse
+from fastapi import FastAPI, File, UploadFile
+import product
 from login import users
 from login.users import database
 from services import check_token
+import cart
+from cart import cart as cart_db
 from shop import notes
 
-app = FastAPI()
+app = FastAPI(title='Zeppelin',
+              description='''Zeppelin is an optical recommendation system powered by the convolutional neural network (CNN) aimed to find the most suitable optical.''',
+              version="1.0.0")
 origins = ["*"]
 
 app.add_middleware(
@@ -72,8 +74,7 @@ async def index(request: Request):
     glasses = await database.fetch_all(query=glasses)
 
     return templates.TemplateResponse("index.html",
-                                      {"request": request, 'glasses': glasses, 'flag': flag,
-                                       'admin': admin})
+                                      {"request": request, 'glasses': glasses, 'flag': flag, 'admin': admin})
 
 
 @app.get('/upload/')
@@ -128,8 +129,7 @@ async def delete_item(request: Request, item_id: str):
     get_id = notes.select().where(notes.c.name == item_id)
     excec_get_id = await database.fetch_one(get_id)
     print(excec_get_id)
-    remove_item = cart_db.delete().where(
-        and_(cart_db.c.user_id == user_id, cart_db.c.glasses_id == excec_get_id[0]))
+    remove_item = cart_db.delete().where(and_(cart_db.c.user_id == user_id, cart_db.c.glasses_id == excec_get_id[0]))
     excex_remove_item = await database.execute(remove_item)
     return 'success'
 
@@ -161,5 +161,5 @@ async def payment(request: Request):
     prices = [i[8] for i in excec_glasses_items]
     total_in_cart = sum(prices)
     return templates.TemplateResponse('payment-page.html',
-                                      {"request": request, 'items': excec_glasses_items,
-                                       'flag': flag, 'sum': total_in_cart})
+                                      {"request": request, 'items': excec_glasses_items, 'flag': flag,
+                                       'sum': total_in_cart})
